test: InputConverter 단위테스트 작성
